feat(viewer): inspect a link's mapping from the links table
Reading what a cross-DC link resolves to meant opening the edit form and
expanding a section inside it. Every row in the Cross-DC links table now
carries a magnifier that opens the inspector directly. It is read-only, so
unlike Edit and Delete it stays enabled in public/demo mode.

The inspector moves out of LinkEditModal into its own module, with a
`collapsible` prop: the edit modal keeps the collapsed Show/Hide section it
had, the standalone modal loads immediately and gets a taller grid.

Three fixes to the grid itself, found while driving it against ampliseq:

- `Via` said "Exact key" for a value that resolves to `S1 - First read`,
  which reads as a claim about the resolved names. It describes the lookup:
  the value IS a mapping key, and the key expands to the canonical ID plus
  every MultiQC variant. Renamed to "Matched via" with a label and a hover
  explanation per rule, and added the four resolver-name vias the endpoint
  emits for non-sample_mapping links, which rendered as bare text before.
- Columns were fixed-width and unsortable-by-menu, and the variant list was
  clipped to one line with no way to widen it. All columns resize, sort and
  filter; "Resolves to" wraps; both value columns carry tooltips; the grid
  paginates at 25 with a page-size selector; cell focus is back for keyboard
  navigation.
- The modal is 90% wide rather than Mantine's 780px `xl`, where the variant
  list wrapped to six lines a row.

The links table keeps its row actions on one line and stops clipping the
widest resolver badge to "sample_map…".
