+++ depictio/viewer/src/projects/detail/LinksSection.tsx
@@ -29,6 +29,7 @@ import {
 import type { DCLink, ResolverInfo } from 'depictio-react-core';
 
 import LinkEditModal from './LinkEditModal';
+import { LinkMappingModal } from './LinkMappingInspector';
 
 interface DataCollectionOption {
   id: string;
@@ -95,6 +96,9 @@ const LinksSection: React.FC<LinksSectionProps> = ({
   // guarantees "Target data collection" (and every other field) starts blank
   // when the user clicks Add link a second time.
   const [modalKey, setModalKey] = useState(0);
+  // Read-only mapping inspection, opened straight from a row's magnifier so
+  // checking what a link resolves to doesn't mean entering the edit form.
+  const [inspecting, setInspecting] = useState<DCLink | null>(null);
   // Collapsed by default to keep the Overview tab compact.
   const [opened, { toggle }] = useDisclosure(false);
   const [filter, setFilter] = useState('');
@@ -219,6 +223,8 @@ const LinksSection: React.FC<LinksSectionProps> = ({
     setModalOpened(true);
   };
 
+  const closeInspect = () => setInspecting(null);
+
   const closeModal = () => {
     setModalOpened(false);
     // Clear `editing` too — without this, a saved edit would leave the
@@ -229,12 +235,16 @@ const LinksSection: React.FC<LinksSectionProps> = ({
   const disabledTip = canMutate ? undefined : 'Disabled in public/demo mode';
 
   return (
-    <Paper withBorder radius="md" p="sm">
+    <Paper withBorder radius="md" p="sm" data-testid="links-section">
       {/* Same header structure as the Data Collections Manager: clickable
           title area, then the primary action, then the chevron as an
           ActionIcon on the far right. */}
       <Group justify="space-between" wrap="nowrap">
-        <UnstyledButton onClick={toggle} style={{ flex: 1, minWidth: 0 }}>
+        <UnstyledButton
+          onClick={toggle}
+          style={{ flex: 1, minWidth: 0 }}
+          data-testid="links-section-toggle"
+        >
           <Group gap="xs" wrap="nowrap">
             <Icon icon="mdi:link-variant" width={20} />
             <Title order={4}>Cross-DC links</Title>
@@ -319,8 +329,15 @@ const LinksSection: React.FC<LinksSectionProps> = ({
                     </Text>
                   </Table.Td>
                   <Table.Td>{tgtTag(link)}</Table.Td>
-                  <Table.Td>
-                    <Badge size="sm" variant="light">
+                  <Table.Td style={{ whiteSpace: 'nowrap' }}>
+                    {/* Badge clips its label by default, which made the widest
+                        resolver name read "sample_map…" once the description
+                        column claimed the slack. */}
+                    <Badge
+                      size="sm"
+                      variant="light"
+                      styles={{ label: { overflow: 'visible' } }}
+                    >
                       {link.link_config?.resolver || 'direct'}
                     </Badge>
                   </Table.Td>
@@ -340,7 +357,18 @@ const LinksSection: React.FC<LinksSectionProps> = ({
                     </Text>
                   </Table.Td>
                   <Table.Td>
-                    <Group gap={4} justify="flex-end">
+                    <Group gap={4} justify="flex-end" wrap="nowrap">
+                      <Tooltip label="Inspect mapping" withArrow>
+                        <ActionIcon
+                          variant="subtle"
+                          color="teal"
+                          onClick={() => setInspecting(link)}
+                          aria-label="Inspect mapping"
+                          data-testid="inspect-link-btn"
+                        >
+                          <Icon icon="mdi:magnify-scan" width={16} />
+                        </ActionIcon>
+                      </Tooltip>
                       <Tooltip label={disabledTip} disabled={canMutate}>
                         <ActionIcon
                           variant="subtle"
@@ -373,6 +401,15 @@ const LinksSection: React.FC<LinksSectionProps> = ({
           )}
         </ScrollArea.Autosize>
       </Collapse>
+
+      <LinkMappingModal
+        opened={!!inspecting}
+        projectId={projectId}
+        link={inspecting}
+        sourceTag={inspecting ? srcTag(inspecting) : ''}
+        targetTag={inspecting ? tgtTag(inspecting) : ''}
+        onClose={closeInspect}
+      />
 
       <LinkEditModal
         key={modalKey}
